chore(meshupgrade): refactor file names
