+++ plugins/lime-plugin-mesh-wide-upgrade/src/containers/nodesList.tsx
@@ -1,7 +1,7 @@
 import Loading from "components/loading";
 
 import NodeUpgradeInfoItem from "plugins/lime-plugin-mesh-wide-upgrade/src/components/nodeUpgradeInfo";
-import { useMeshUpgrade } from "plugins/lime-plugin-mesh-wide-upgrade/src/hooks/MeshWideUpgradeProvider";
+import { useMeshUpgrade } from "plugins/lime-plugin-mesh-wide-upgrade/src/hooks/meshWideUpgradeProvider";
 
 export const NodesList = () => {
     const { data, isLoading } = useMeshUpgrade();
