+++ plugins/lime-plugin-mesh-wide-upgrade/src/meshUpgradeTypes.ts
